+++ MeshLib/Elements/Cell.h
@@ -22,7 +22,7 @@ public:
 	virtual double getVolume() const { return _volume; };
 
 	/// Get dimension of the mesh element.
-	size_t getDimension() const { return 3; };
+	unsigned getDimension() const { return 3; };
 
 	/// Destructor
 	virtual ~Cell();
@@ -30,10 +30,10 @@ public:
 protected:
 /*
 	/// Constructor for a generic mesh element containing an array of mesh nodes.
-	Cell(Node** nodes, MshElemType::type type, size_t value = 0);
+	Cell(Node** nodes, MshElemType::type type, unsigned value = 0);
 */
 	/// Constructor for a generic mesh element without an array of mesh nodes.
-	Cell(MshElemType::type type, size_t value = 0);
+	Cell(MshElemType::type type, unsigned value = 0);
 
 	/// Calculate the volume of this 3d element.
 	virtual double calcVolume() = 0;
@@ -42,7 +42,7 @@ protected:
 
 }; /* class */
 
-} /* namespace */
+}		
 
 #endif /* CELL_H_ */
