+++ MeshLib/Elements/Element.cpp
@@ -13,12 +13,12 @@
 namespace MeshLib {
 
 /*
-Element::Element(Node** nodes, MshElemType::type type, size_t value)
+Element::Element(Node** nodes, MshElemType::type type, unsigned value)
 	: _nodes(nodes), _type(type), _value(value)
 {
 }
 */
-Element::Element(MshElemType::type type, size_t value)
+Element::Element(MshElemType::type type, unsigned value)
 	: _nodes(NULL), _type(type), _value(value)
 {
 }
@@ -29,20 +29,20 @@ Element::~Element()
 	delete[] this->_neighbors;
 }
 
-const Element* Element::getNeighbor(size_t i) const
+const Element* Element::getNeighbor(unsigned i) const
 {
 	assert(i < getNNeighbors() && "Error in MeshLib::Element::getNeighbor() - Index does not exist.");
 	return _neighbors[i];
 }
 
-const Node* Element::getNode(size_t i) const
+const Node* Element::getNode(unsigned i) const
 {
 	assert(i < getNNodes() && "Error in MeshLib::Element::getNode() - Index does not exist.");
 	assert(_nodes[i] != NULL && "Error in MeshLib::Element::getNode() - Node is NULL.");
 	return _nodes[i];
 }
 
-size_t Element::getNodeIndex(size_t i) const 
+unsigned Element::getNodeIndex(unsigned i) const 
 {
 	assert(i<getNNodes() && "Error in MeshLib::Element::getNodeIndex() - Index does not exist.");
 	return _nodes[i]->getID();
@@ -50,8 +50,8 @@ size_t Element::getNodeIndex(size_t i) const
 
 bool Element::hasNeighbor(Element* elem) const
 {
-	size_t nNeighbors (this->getNNeighbors());
-	for (size_t i=0; i<nNeighbors; i++)
+	unsigned nNeighbors (this->getNNeighbors());
+	for (unsigned i=0; i<nNeighbors; i++)
 		if (this->_neighbors[i]==elem)
 			return true;
 	return false;
