+++ MeshLib/FemNode.cpp
@@ -9,12 +9,12 @@
 
 namespace MeshLib {
 
-FemNode::FemNode(double const*const coords, size_t id)
+FemNode::FemNode(double const*const coords, unsigned id)
 	: Node(coords, id)
 {
 }
 
-FemNode::FemNode(double x, double y, double z, size_t id)
+FemNode::FemNode(double x, double y, double z, unsigned id)
 	: Node(x, y, z, id)
 {
 }
